Revert "Use Java 8 standards, up-to-date Spring standards, and AssertJ instea..."

+++ src/main/java/com/exo/ecommerce/domain/invoice/Invoice.java
@@ -36,7 +36,8 @@ public class Invoice {
         return cart;
     }
 
-    public double calculateTotalPrice() {
+    public float calculateTotalPrice()
+    {
         return this.cart.calculateTotalPrice();
     }
