+++ src/main/java/com/exo/ecommerce/infrastructure/http/presentation/CurrencyFormatter.java
@@ -8,7 +8,8 @@ import java.util.Locale;
 public class CurrencyFormatter {
     private static DecimalFormat numberFormatInstance = null;
 
-    public static String formatCurrency(double amount) {
+    public static String formatCurrency(float amount)
+    {
         if (numberFormatInstance == null) {
             DecimalFormat numberFormat = (DecimalFormat) DecimalFormat.getCurrencyInstance(Locale.FRANCE);
             numberFormat.applyPattern("###,###.## ¤");
